tg_bot_backend/backend_requirements.txt
anyio==4.4.0
async-property==0.2.2
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
deprecation==2.1.0
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-keycloak==4.1.0
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
